requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
pydantic==2.11.4
langchain==0.3.25
langchain-core==0.3.59
langchain-google-genai==2.1.4
google-generativeai==0.8.5
faiss-cpu==1.8.0.post1
numpy==1.26.4
requests==2.32.3
beautifulsoup4==4.12.3
httpx==0.28.1
